- Package name fixed - Travis.ci support added - Codecov.io support added
