+++ .travis.yml
@@ -1,0 +1,7 @@
+language: java
+
+jdk:
+  - openjdk8
+
+after_success:
+  - bash <(curl -s https://codecov.io/bash)
