--- src/main/java/com/vendorbear/WebApplication.java
@@ -1,11 +1,0 @@
-package com.vendorbear;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WebApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(WebApplication.class, args);
-	}
-}
