--- src/main/java/com/vendorbear/domain/card/exception/CardNotFoundException.java
@@ -1,7 +1,0 @@
-package com.vendorbear.domain.card.exception;
-
-public class CardNotFoundException extends Exception {
-    public CardNotFoundException(String cardReference) {
-        super("Card not found with given reference:" + cardReference);
-    }
-}
