--- src/main/java/com/vendorbear/entitiy/Card.java
@@ -1,127 +1,0 @@
-package com.vendorbear.entitiy;
-
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.Where;
-
-import javax.persistence.*;
-import java.sql.Date;
-
-@Entity
-@Table(name = "card")
-@Where(clause = "ISNULL(is_deleted) OR is_deleted = 0")
-@SQLDelete(sql = "UPDATE card SET is_deleted = 1 WHERE id = ?")
-public class Card {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private long id;
-
-    @Column(name = "balance", nullable = false)
-    private long balance;
-
-    @ManyToOne(targetEntity = Currency.class, fetch = FetchType.EAGER)
-    @JoinColumn(name = "currency_id", nullable = false)
-    private Currency currency;
-
-    @Column(name = "activation_date", nullable = false)
-    private Date activationDate;
-
-    @Column(name = "expire_date", nullable = false)
-    private Date expireDate;
-
-    @Column(name = "reference", nullable = false)
-    private String reference;
-
-    @Column(name = "card_number", nullable = false)
-    private String cardNumber;
-
-    @Column(name = "cvc", nullable = false)
-    private String cvc;
-
-    @Column(name = "active", nullable = false)
-    private boolean active = true;
-
-    @Column(name = "is_deleted")
-    private boolean isDeleted = false;
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public long getBalance() {
-        return balance;
-    }
-
-    public void setBalance(long balance) {
-        this.balance = balance;
-    }
-
-    public Currency getCurrency() {
-        return currency;
-    }
-
-    public void setCurrency(Currency currency) {
-        this.currency = currency;
-    }
-
-    public Date getActivationDate() {
-        return activationDate;
-    }
-
-    public void setActivationDate(Date activationDate) {
-        this.activationDate = activationDate;
-    }
-
-    public Date getExpireDate() {
-        return expireDate;
-    }
-
-    public void setExpireDate(Date expireDate) {
-        this.expireDate = expireDate;
-    }
-
-    public String getReference() {
-        return reference;
-    }
-
-    public void setReference(String reference) {
-        this.reference = reference;
-    }
-
-    public String getCardNumber() {
-        return cardNumber;
-    }
-
-    public void setCardNumber(String cardNumber) {
-        this.cardNumber = cardNumber;
-    }
-
-    public String getCvc() {
-        return cvc;
-    }
-
-    public void setCvc(String cvc) {
-        this.cvc = cvc;
-    }
-
-    public boolean isActive() {
-        return active;
-    }
-
-    public void setActive(boolean active) {
-        this.active = active;
-    }
-
-    public boolean isDeleted() {
-        return isDeleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        isDeleted = deleted;
-    }
-}
