--- src/main/java/com/vendorbear/schema/Card.java
@@ -1,69 +1,0 @@
-package com.vendorbear.schema;
-
-import java.util.Date;
-
-public class Card {
-    private long balance;
-    private String currency;
-    private Date activationDate;
-    private Date expireDate;
-    private String reference;
-    private String cardNumber;
-    private String cvc;
-
-    public long getBalance() {
-        return balance;
-    }
-
-    public void setBalance(long balance) {
-        this.balance = balance;
-    }
-
-    public String getCurrency() {
-        return currency;
-    }
-
-    public void setCurrency(String currency) {
-        this.currency = currency;
-    }
-
-    public Date getActivationDate() {
-        return activationDate;
-    }
-
-    public void setActivationDate(Date activationDate) {
-        this.activationDate = activationDate;
-    }
-
-    public Date getExpireDate() {
-        return expireDate;
-    }
-
-    public void setExpireDate(Date expireDate) {
-        this.expireDate = expireDate;
-    }
-
-    public String getReference() {
-        return reference;
-    }
-
-    public void setReference(String reference) {
-        this.reference = reference;
-    }
-
-    public String getCardNumber() {
-        return cardNumber;
-    }
-
-    public void setCardNumber(String cardNumber) {
-        this.cardNumber = cardNumber;
-    }
-
-    public String getCvc() {
-        return cvc;
-    }
-
-    public void setCvc(String cvc) {
-        this.cvc = cvc;
-    }
-}
