+++ src/main/java/com/vendorlion/WebApplication.java
@@ -1,0 +1,11 @@
+package com.vendorlion;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WebApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(WebApplication.class, args);
+	}
+}
