+++ src/main/java/com/vendorlion/domain/card/AbstractCardService.java
@@ -1,0 +1,13 @@
+package com.vendorlion.domain.card;
+
+import com.vendorlion.repository.CardRepository;
+import org.springframework.stereotype.Service;
+
+@Service
+public abstract class AbstractCardService {
+    protected final CardRepository cardRepository;
+
+    public AbstractCardService(CardRepository cardRepository) {
+        this.cardRepository = cardRepository;
+    }
+}
