+++ src/main/java/com/vendorlion/domain/card/exception/CardNotFoundException.java
@@ -1,0 +1,7 @@
+package com.vendorlion.domain.card.exception;
+
+public class CardNotFoundException extends Exception {
+    public CardNotFoundException(String cardReference) {
+        super("Card not found with given reference:" + cardReference);
+    }
+}
