+++ src/main/java/com/vendorlion/entitiy/Currency.java
@@ -1,0 +1,32 @@
+package com.vendorlion.entitiy;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "currency")
+public class Currency {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id")
+    private long id;
+
+    @Column(name = "code")
+    private String code;
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String code) {
+        this.code = code;
+    }
+}
