+++ src/main/java/com/vendorlion/web/api/security/APISecurityConfigurerAdapter.java
@@ -1,0 +1,31 @@
+package com.vendorlion.web.api.security;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+public class APISecurityConfigurerAdapter extends WebSecurityConfigurerAdapter
+{
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+            .csrf().disable()
+            .authorizeRequests().anyRequest().authenticated()
+            .and()
+            .httpBasic()
+        ;
+    }
+
+    @Autowired
+    public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
+        // TODO: Fetch authentication users from database
+        auth.inMemoryAuthentication()
+            .withUser("admin")
+            .password("{noop}password")
+            .roles("USER")
+        ;
+    }
+}
